Unreal_Wwise_Starting_Troubleshooting
When pressed play, everything is played at once, including events that are not in the area. Blueprints were used and post event with set occlusion is set. In to check for landscape collision, spatial audio volume and see why it is happening. In addition, realizes there might be an explanation as well when it is in relation to player (self). As there is no first person player blueprint or fps game mode. Will continue and check with Andre, and online for solutions.